non_empty_string

Struct NonEmptyString

pub struct NonEmptyString(/* private fields */);

A simple String wrapper type, similar to NonZeroUsize and friends. Guarantees that the String contained inside is not of length 0.

Implementations

impl NonEmptyString

pub fn new(string: String) -> Result<Self, String>

Attempts to create a new NonEmptyString. If the given string is empty, Err is returned, containing the original String, Ok otherwise.

pub unsafe fn new_unchecked(string: String) -> Self

Creates a new NonEmptyString, assuming string is not empty.

Safety

If the given string is empty, it’ll be undefined behavior.

pub fn get(&self) -> &str

Returns a reference to the contained value.

pub fn into_inner(self) -> String

Consume the NonEmptyString to get the internal String out.

pub fn into_bytes(self) -> Vec<u8>

Is forwarded to the inner String. See String::into_bytes

pub fn as_str(&self) -> &str

Is forwarded to the inner String. See String::as_str

pub fn push_str(&mut self, string: &str)

Is forwarded to the inner String. See String::push_str

pub fn capacity(&self) -> usize

Is forwarded to the inner String. See String::capacity

pub fn reserve(&mut self, additional: usize)

Is forwarded to the inner String. See String::reserve

pub fn reserve_exact(&mut self, additional: usize)

Is forwarded to the inner String. See String::reserve_exact

pub fn try_reserve_exact( &mut self, additional: usize, ) -> Result<(), TryReserveError>

Is forwarded to the inner String. See String::try_reserve_exact

pub fn shrink_to_fit(&mut self)

Is forwarded to the inner String. See std::string::String::[(&]

pub fn shrink_to(&mut self, min_capacity: usize)

Is forwarded to the inner String. See String::shrink_to

pub fn push(&mut self, ch: char)

Is forwarded to the inner String. See String::push

pub fn as_bytes(&self) -> &[u8]

Is forwarded to the inner String. See String::as_bytes

pub fn insert(&mut self, idx: usize, ch: char)

Is forwarded to the inner String. See String::insert

pub fn insert_str(&mut self, idx: usize, string: &str)

Is forwarded to the inner String. See String::insert_str

pub fn len(&self) -> usize

Is forwarded to the inner String. See String::len

pub fn into_boxed_str(self) -> Box<str>

Is forwarded to the inner String. See String::into_boxed_str